+++ database/migrations/2020_03_31_145421_alter_table_action_instances_add_should_queue_column.php
@@ -1,0 +1,32 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class AlterTableActionInstancesAddShouldQueueColumn extends Migration
+{
+    /**
+     * Run the migrations.
+     * 
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('action_instances', function (Blueprint $table) {
+            $table->boolean('should_queue')->default(true);
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('action_instances', function (Blueprint $table) {
+            $table->dropColumn('should_queue');
+        });    
+    }
+}